UNSTABLE f(n) function finished
i dont know whats going on but i assume when there are two distances that are the same, heapq tries to compare the Node instead of the distance (read comments in the Node class). I tried to fix that by making it so that the more populous place/village get prioritised. Sadly, this yields weird results from the algorithm as it will sometimes yield "inf" for f(n) and sometimes, it would work. I can't point out the underlying issue with that. My assumption is still that heapq tries to compare Nodes even if it is supposed to compare the integer distances in the tuple